GUI: Show prompt dialog if closing without changes
When the shell widget was closed a call to ':qa' was made using
vim_command().

But vim_command does not display error messages, and :qa prints an error
msg and blocks waiting for user input. In our case no messages was being
displayed, but the user was still prompted to press a key.

This commit uses ':confirm qa' instead of simply, ':qa'. The user is
then prompted to save/discard/cancel for each buffer.

Fixes #310.